improve test case and fix bug

+++ src/main/scala/com/capslock/leveldb/SequenceNumber.scala
@@ -17,6 +17,6 @@ object SequenceNumber {
     }
 
     def unpackValueType(packedData: Long): ValueType = {
-        ValueType(packedData.toInt)
+        ValueType(packedData.toByte)
     }
 }
